Fix encoding mismatch in shell
Change ssh shell to explicitly encode strings in utf-8 and use
the length of the byte string instead of the length of a string
when checking if all bytes were sent through the shell ssh channel.

Always encode and decode shell channel data in utf-8.

Check if the locale charmap of the target host is UTF-8 and issue
a warning otherwise.