Make sure to turn on unlockDelay in s3270 cmdline call
In newer versions of s3270 the default value of the 'unlockDelay'
resource changed from on to off which causes sporadic problems
with the zvm terminal releasing the console too early when the
keyboard is still locked leading to failure as the next command
is ignored. Therefore we explicitly specify the option in the
command line call to make sure it's always on.