Adds loop for defining cpus for cms guests to limit output buffer
When defining a large number of cpus (16+) for a zVM guest a user
may encounter an exception with the 'HCPLGA054E: Already logged on'
message and have to manually IPL the guest after the session fails.

A large enough range of cpu addresses can produce enough output to
require clearing the screen to continue. The regular expression
being used to scrape the output looks for a generic success rather
than the success of a specific address. This results in the existing
screen clearing functionality being unaware that there is additional
output to handle because a match was found in the current buffer.

This change adds a while loop around the define cpu command to
limit the amount of output produced by each iteration.

The alternative to a loop was modifying the regular expression
to look for the last cpu address that was defined. Thought it safer
to not assume the output would always appear in the same order.

DCO 1.1 Signed-off-by: Matt Gritter <mrgritte@us.ibm.com>